+++ modape/scripts/modis_download.py
@@ -59,8 +59,10 @@ def main():
 
     credentials = Credentials(args.username, args.password)
 
+    if args.aria2:
+        warnings.warn("\nARIA2 is now standard download tool. The flag is therefore depracated and will raise an error in a future release!\n Please consider removing it,", DeprecationWarning)
+
     if args.download:
-        warnings.warn("Download only possible with ARIA2! Flag --aria2 will be removed in future release.", DeprecationWarning)
 
         #fail if ARIA2 not installed
         try: